feat: Add removeValueByKey and addValueByKey functions to helper.ts
The code changes introduce two new functions, `removeValueByKey` and `addValueByKey`, in the `helper.ts` file. These functions allow the App component to remove and add values to specific keys in an array of objects based on specified criteria. This enhances the functionality of the application by providing more flexibility in manipulating object properties.

Resolves: #123